refactor: move activation of vs code test commands for aws.q section to one file
