+++ .vscode/launch.json
@@ -88,8 +88,7 @@
                 "ENABLE_TOKEN_PROVIDER": "true",
                 "ENABLE_CUSTOM_COMMANDS": "true",
                 "ENABLE_CHAT": "true",
-                "ENABLE_CUSTOMIZATIONS": "true",
-                "ENABLE_DEVELOPER_PROFILES": "true"
+                "ENABLE_AWS_Q_SECTION": "true"
                 // "HTTPS_PROXY": "http://127.0.0.1:8888",
                 // "AWS_CA_BUNDLE": "/path/to/cert.pem"
             },
