+++ client/vscode/package.json
@@ -88,6 +88,11 @@
                 "category": "Test Amazon Q Extension"
             },
             {
+                "command": "aws.sample-vscode-ext-amazonq.getAwsQSection",
+                "title": "Get entire AWS Q section",
+                "category": "Test Amazon Q Extension"
+            },
+            {
                 "command": "aws.sample-vscode-ext-amazonq.getCustomizations",
                 "title": "Get customizations",
                 "category": "Test Amazon Q Extension"
