+++ client/vscode/src/activation.ts
@@ -20,10 +20,9 @@ import {
 import { registerInlineCompletion } from './inlineCompletionActivation'
 import { registerLogCommand, registerTransformCommand } from './sampleCommandActivation'
 import { randomUUID } from 'crypto'
-import { registerCustomizations } from './customizationActivation'
 import { registerIdentity } from './identityActivation'
 import { registerNotification } from './notificationActivation'
-import { registerDeveloperProfiles } from './developerProfilesActivation'
+import { registerAwsQSection } from './awsQSectionActivation'
 
 export async function activateDocumentsLanguageServer(extensionContext: ExtensionContext) {
     /**
@@ -196,14 +195,9 @@ export async function activateDocumentsLanguageServer(extensionContext: Extensio
         registerChat(client, extensionContext.extensionUri, enableEncryptionInit ? encryptionKey : undefined)
     }
 
-    const enableCustomizations = process.env.ENABLE_CUSTOMIZATIONS === 'true'
-    if (enableCustomizations) {
-        registerCustomizations(client)
-    }
-
-    const enableDeveloperProfiles = process.env.ENABLE_DEVELOPER_PROFILES === 'true'
-    if (enableDeveloperProfiles) {
-        registerDeveloperProfiles(client)
+    const enableAwsQSection = process.env.ENABLE_AWS_Q_SECTION === 'true'
+    if (enableAwsQSection) {
+        registerAwsQSection(client)
     }
 
     const enableIdentity = process.env.ENABLE_IDENTITY === 'true'
